统一db 带宽，内存相关单位统一为B 和B/s

Created-by: msprofDeveloper

Match-id-fb07e5df15e7bc2e9bb1d212847659abdcc9743b

+++ test/msprof_cpp/analysis_ut/viewer/test/ddr_processor_utest.cpp
@@ -120,7 +120,7 @@ TEST_F(DDRProcessorUTest, TestRunShouldReturnFalseWhenInsertDataFailed)
 
 TEST_F(DDRProcessorUTest, TestRunShouldReturnFalseWhenReserveFailedThenDataIsEmpty)
 {
-    using TempT = std::tuple<uint16_t, uint64_t, double, double>;
+    using TempT = std::tuple<uint16_t, uint64_t, uint64_t, uint64_t>;
     MOCKER_CPP(&std::vector<TempT>::reserve).stubs().will(throws(std::bad_alloc()));
     auto processor = DDRProcessor(DB_PATH, PROF_PATHS);
     EXPECT_FALSE(processor.Run());
